requirements.txt
requests
lockfile
psutil
python-daemon
pyparsing
cifparser
setuptools
